Fix missed deletions
Logs from blocks where the hash begins with a leading zero were not
being deleted in the event of a reorg because we were not trimming
the order hash as expected by the database.

While we were troubleshooting that, we realized that having buffered
channels for added and removed logs could result in added and removed
logs being processed out of order, so a log that was supposed to be
added then removed might errantly be removed then added, so we also
made those unbuffered channels.