Holders API, pagination
This adds a flume_erc20Holders API, that will return the token holders
for an ERC20 token contract. It's comparable to flume_erc20ByAccount,
but mapping tokens to holders instead of holders to their tokens.

Both of these APIs are paginated. If the response set exceeds 10,000
items it will include a continuation token, which can be passed a a
second parameter.