Switch to LIMIT / OFFSET for pagination
Doing pagination based on a field that wasn't the index we're searching
on meant that sqlite had to gather the entire dataset before it could
sort the results, so it wasn't significantly faster than just returning
the entire result set. Using LIMIT / OFFSET is a bit more prone to
missing stuff / duplication as the dataset changes, but makes the
performance much more predictable.