Use chain feed to index transactions and logs
This transitions from a feed of just logs to a feed of blocks, transactions,
and logs to index the database.

Currently we only have a websockets implementation (no Kafka) and the
request handler hasn't been updated for the schema changes (though for
existing queries that should just be a matter of repointing them to the
v_event_logs view instead of the event_logs table.

Schema changes:

* Transactions are now stored in the database
* Transaction hashes, block numbers, and other transaction data from
  the logs has been moved to the transaction table (removing it from
  the logs table), and a foreign key has been established to link the
  two.

Indexing changes:

* The indexer now works block by block, getting notifications of new blocks
  from a websockets subscription, then getting the transaction details
  and receipts (including logs) with successive queries.
* Absent a deletion notification, when a new block comes in we drop any transactions
  (which then cascades to drop logs) of a block at the same height with a
  different hash.