Make Flume's Kafka resumption topic aware
Prior to now, resuming from Kafka after migrating to a different
topic or broker meant you had to go in and manually remove the
offsets. Now, the offset is stored in conjunction with the topic,
so if the topic name changes it will pull the oldest value from
the topic.