Initial work on mempool in flume
This creates an in-memory sqlite database to track the mempool, and
an indexer to feed data into it. By default, a nil feed will be used,
so the mempool will be empty.

Pending transactions will continue to be tracked for a while after they
are confirmed, with the intention of using this to estimate confirmation
durations for upcoming transactions.