Optimize db inserts
This optimizes database inserts by compiling one big statement with
all the inserts for the block, and passing all the parameters into
that statement at once.

The goal is to avoid crossing the cgo barrier more times than necessary,
as that has a non-negligible cost associated with it. If crossing the cgo
boundary takes 0.5ms each time, then a block with 100txs and 500 logs takes
a minimum of 300ms just for cgo calls.