Try parallelizing transaction processing
I'm trying to see if we can eek out a little bit better block processing
time by processing each transaction in parallel. The statements are put
onto channels, and the channels are processed in order, so the same
SQL statements should end up being executed, but they'll be compiled
in parallel.