Use block td in resumption
Recent changes to the upstream chain event feed require the block
total difficulty during the resumption process. I realized as I was
implementing the flume side of this that we were inadvertently
truncating the total difficulty stored in flume to 64 bits (when it
should be a 256 bit value). This has been corrected in the code, and
a migration of our existing (pre-production) database is in progress.