Denormalize transaction and block hashes into logs
We've been seeing significant performance issues with log queries
that we believe will be resolved by denormalizing transaction and block
hashes into the event_logs table.

In the old version of flume (logs only) this data all lived on the same
table, and the read operations for production flume are quite low. With
the new version of flume where transaction and block hashes must be
retrieved from the transaction and block tables, the number of read operations
were substantially higher, despite well-indexed operations.

The hope is that denormalizing this data will reduce the number of disk read
operations back to the levels we saw with classic flume, with a relatively
minor penalty in terms of disk usage.