Reuse the same snapshot volume
A significant portion of the time spent taking snapshots is in
pulling down the snapshot from s3. This change will create a volume
for snapshotting if none exists in the current AZ, then on subsequent
snapshots will look for an existing volume and mount it. This should
reduce snapshot times considerably.

A couple of notes for future work:

* It might make sense to make st1 or sc1 volume types an option
  for snapshots. They're slower, but probably not enough slower
  to overcome not having to download the database, and they're
  enough cheaper that if we're going to keep them around all
  the time we probably want the savings.
* We'll want to make sure that the snapshotter process will give
  preference to the same availability zone. Right now it gives
  preference by instance types, cycling through AZs to try and
  find one that can provide the desired instance type. If we have
  volumes lying around, we'd rather cycle through instance types
  to find one in the AZ that has the volume.