Use updated snapshotter
This uses the updated snapshotter which will sort by network first,
then use the instance type overrides specified here. This will let us
get snapshots with arm instance types, and better take advantage of
pre-existing snapshot volumes.

Also, it rearranges the AZs to try, because it seems us-east-2
does not have instances in AZ 'a', so rather than try and fail
with that AZ, start with one that exists.