Limit snapshots to one AZ
Now that volumes are being retained, if a snapshot happens in the
non-primary AZ, it will create a new volume. If that volume doesn't
get used again during the Kafka retention period, it won't be valid
for making a snapshot on the next attempt, resulting in failures.

Pinning to a single AZ does mean that attempt to create a snapshot
will fail, but we'll be alerted to that by the garbage collector
and can retrigger later.