Don't filter by block range and block hash in getLogs
Specifying both a block range and block hash is invalid in getLogs,
but if a block range is not specified, it was implicitly from latest
to latest. Unless the blockhash was the latest block's hash, this meant
that the range portion of the query and the block hash portion would
be mutually exclusive, and results would be empty.

Now we have the problem that blockHash is not indexed on the event_logs
table. This would mean if someone only specified a blockHash, Flume
would have to scan the entire event_logs table, which would time out
with production data. We could join against the blocks table, but then
we'd be doing a join for all getLogs queries when it's needed for only
a small set of them. Instead, if blockHash is specified we're querying
for the block number, and adding that to our query so we have an
opportunity for an efficient index.