Ensure logs are sorted
Logs should be sorted by block, then the log index within the block.

Scanning the logs table ensures this, but some indexes do not.

Additionally, we don't want to use an "ORDER BY" statement, as that
will require getting the whole data set and sorting it before we start
processing the results, and it is while processing results that we
check that result sets do not exceed 10k entries.