Initial work on schema refactor
See migrations/schema.sql for details on proposed DB schema.

Introduces a concept of cardinal-streams based indexers to flume,
having independent indexers provide SQL statements to be committed
with each chain update. This will allow us to enable and disable
indexers along with sets of features we want a flume instance to support.